// backend/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  VIEWER: 'VIEWER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ContractStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  CLOSED: 'CLOSED'
} as const

export type ContractStatus = (typeof ContractStatus)[keyof typeof ContractStatus]


export const HistoryAction = {
  CREATE: 'CREATE',
  UPDATE_FIELD: 'UPDATE_FIELD',
  STATUS_CHANGE: 'STATUS_CHANGE'
} as const

export type HistoryAction = (typeof HistoryAction)[keyof typeof HistoryAction]